requirements.txt
pipenv
pymongo
datetime
python-dateutil
flask
flask-cors
gunicorn
certbot
certbot-nginx
sense-energy
dotenv
